% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta_select}
\alias{meta_select}
\title{Method for selecting meta-parameters for \code{swag}}
\usage{
meta_select(x, y, space_exp = 0.5, pmax = 3, caret_args_dyn = NULL, ...)
}
\arguments{
\item{x}{A \code{matrix} or \code{data.frame} of attributes}

\item{y}{A \code{vector} of binary response variable.}

\item{space_exp}{A \code{double} representing the percentage of the space of all possible learners of size two to explore.}

\item{pmax}{A \code{integer} representing the maximum number of attributes per learner.}

\item{caret_args_dyn}{If not \code{NULL}, a function that can modify arguments for
\code{\link[caret]{train}} dynamically (see 'Details').}
}
\value{
\code{meta_select} returns a \code{list} with the following components:
\tabular{ll}{
   \code{control} \tab the \code{control} that contains the suggested \code{swag} meta-parameters (see \code{\link{swagControl}}) \cr
   \code{max_eta} \tab a \code{vector} containing the \code{swag} maximum expected time of arrival (eta) for each dimension up to \code{pmax}.
   The eta is obtained directly from \code{\link[microbenchmark]{microbenchmark}} function with a default value of 5 repetitions. \cr
}
}
\description{
The Sparse Wrapper AlGorithm depends on some meta-parameters that are
 selected by this function given the learning ingredients described below.
}
\details{
Currently we expect the user to replace \code{...} with the arguments one would
use for \code{\link[caret]{train}}. This requires to know how to use \code{\link[caret]{train}}
function. If \code{...} is left unspecified, default values of \code{\link[caret]{train}}
are used. But this might lead to unexpected results.

The function \code{caret_args_dyn} is expected to take as a first
argument a \code{list} with all arguments for \code{\link[caret]{train}}
and as a second argument the number of attributes (see examples in the vignette).

More specifically, \code{meta_select} determines the meta-parameters that are necessary to
train the Sparse Wrapper AlGorithm thanks to the choice of \code{space_exp}. Indeed this
selection enables the user to choose the preferred percentage of the space of all possible
learners of size two to explore. See \insertCite{molinari2020swag;textual}{swag} for more details.
Furthermore, since a given combination of meta-parameters determines automatically a maximum
computational time, we provide a maximum expected time of arrival (ETA) for each dimension up
to the \code{pmax} selected by the user.
}
\author{
Cesare Miglioli and Samuel Orso
}
