% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swag.R
\name{print.swag}
\alias{print.swag}
\title{Print swag object}
\usage{
\method{print}{swag}(x, ...)
}
\arguments{
\item{x}{A \code{swag} object}

\item{...}{Other arguments passed to specific methods}
}
\value{
Text output via print
}
\description{
Displays information about swag object
}
\author{
Gaetan Bakalli
}
\keyword{internal}
