% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{return_lm_beta_selected_models}
\alias{return_lm_beta_selected_models}
\title{Returned estimated linear regression coefficients for each selected model in a summary.swag object}
\usage{
return_lm_beta_selected_models(swag_summary)
}
\arguments{
\item{swag_summary}{A \code{swag_summary} object.}
}
\description{
The function return a list that contains beta_models_df and the swag_summary object.
The beta_models_df is a dataframe where the columns are all of the selected variables from the summary.swag objects,
and where each row are the estimated coefficients from a selected models using the classical lm procedure.
}
\author{
Gaetan Bakalli, Samuel Orso, Cesare Miglioli and Lionel Voirol
}
