% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seer.R
\name{seer}
\alias{seer}
\title{SEER wrapper algorithm for ML method.}
\usage{
seer(
  y,
  X,
  learner = "logistic",
  dmax = NULL,
  m = NULL,
  q0 = 0.01,
  seed = 163,
  parallel_comput = T,
  nc = NULL,
  ...
)
}
\arguments{
\item{y}{A \code{vector} of binary response variable.}

\item{X}{A \code{matrix} or \code{data.frame} of attributes}

\item{learner}{A \code{string} defining the learner type (method available: \code{"logistic"}, \code{"svmLinear"},
\code{"svmRadial"}, \code{"lasso"} and \code{"rf"})}

\item{dmax}{A \code{double} representing the maximum number of attributes per learner.}

\item{m}{A \code{integer} representing the maximum number of learners per dimension explored.}

\item{q0}{A \code{double} representing the proportion of screening.}

\item{seed}{An \code{integer} that controls the reproducibility.}

\item{parallel_comput}{An \code{boolean} to allow for parallel computing.}

\item{nc}{An \code{double} that specify the number of core for parallel computation.}
}
\value{
A \code{seer} object with the structure:
\describe{
\item{}{}
}
}
\description{
SEER algo
}
\examples{
seer()
}
\author{
Gaetan Bakalli and Samuel Orso
}
