% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seer.R
\name{seer_valid}
\alias{seer_valid}
\title{SEER validation for ML method.}
\usage{
seer_valid(obj, y_valid, X_valid, seed = 163)
}
\arguments{
\item{obj}{A \code{object} of of class \code{'seer'}.}

\item{seed}{An \code{integer} that controls the reproducibility.}

\item{X}{A \code{matrix} or \code{data.frame} of attributes}

\item{parallel_comput}{An \code{boolean} to allow for parallel computing.}

\item{nc}{An \code{double} that specify the number of core for parallel computation.}
}
\value{
A \code{seer} object with the structure:
\describe{
\item{}{}
}
}
\description{
SEER algo
}
\examples{
seer()
}
\author{
Gaetan Bakalli and Samuel Orso
}
