% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swag.R
\name{swag}
\alias{swag}
\title{swag wrapper algorithm for ML method.}
\usage{
swag(x, y, control = swagControl(), auto_control = TRUE, ...)
}
\arguments{
\item{x}{A \code{matrix} or \code{data.frame} of attributes}

\item{y}{A \code{vector} of binary response variable.}

\item{control}{see \code{swagControl()}}

\item{auto_control}{A \code{boolean}, whether some control parameters
are adjusted depending on `x` and `y` (see \code{swagControl()}).}
}
\value{
A \code{swag} object with the structure:
\describe{
\item{}{}
}
}
\description{
swag algo
}
\examples{
swag()
}
\author{
Gaetan Bakalli, Samuel Orso and Cesare Miglioli
}
