% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swag.R
\name{swag_valid}
\alias{swag_valid}
\title{swag validation for ML method.}
\usage{
swag_valid(obj, y_valid, X_valid, seed = 163)
}
\arguments{
\item{obj}{A \code{object} of of class \code{'swag'}.}

\item{seed}{An \code{integer} that controls the reproducibility.}

\item{X}{A \code{matrix} or \code{data.frame} of attributes}

\item{parallel_comput}{An \code{boolean} to allow for parallel computing.}

\item{nc}{An \code{double} that specify the number of core for parallel computation.}
}
\value{
A \code{swag} object with the structure:
\describe{
\item{}{}
}
}
\description{
swag algo
}
\examples{
swag()
}
\author{
Gaetan Bakalli and Samuel Orso
}
