% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{swagControl}
\alias{swagControl}
\title{Control for swag function}
\usage{
swagControl(
  pmax = 3,
  m = 100,
  alpha = 0.05,
  seed = 163L,
  verbose = FALSE,
  verbose_dim_1 = FALSE
)
}
\arguments{
\item{pmax}{A \code{integer} representing the maximum number of attributes per learner.}

\item{m}{A \code{integer} representing the maximum number of learners per dimension explored.}

\item{alpha}{A \code{double} representing the proportion of screening.}

\item{seed}{An \code{integer} that controls the reproducibility.}

\item{verbose}{A \code{boolean} for printing current progress of the algorithm.}
}
\description{
The Spare Wrapper AlGorithm depends on some meta-parameters that are
described below.
}
\seealso{
\code{\link[swag]{swag}}
}
