% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.swag}
\alias{summary.swag}
\title{Summary method for SWAG}
\usage{
\method{summary}{swag}(
  object,
  min_dim_method = "median",
  min_dim_min_cv_error_quantile = 0.01,
  ...
)
}
\arguments{
\item{object}{A \code{object}.}

\item{min_dim_method}{A \code{string} that specify the method to identify the dimension on which to compute the quantile to set the minimal CV to select model.}

\item{min_dim_min_cv_error_quantile}{The quantile of CV error in the selected dimension to specify the minimum CV value for selected models.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Method `summary` that returns the number and proportion of appearance of each variables on a subset of selected model.
The selection procedure of models proceed in two steps. First we select an explored dimension in which the `mean`, `min` or `median` is the lowest.
 We then compute the selected percentile of the CV error on this dimension.
 We then select all models in all explored dimensions that have a lower CV error than the CV value set by this two-steps procedure.
}
\author{
Gaetan Bakalli, Samuel Orso, Cesare Miglioli and Lionel Voirol
}
