% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.swag}
\alias{summary.swag}
\title{Summary method for SWAG}
\usage{
\method{summary}{swag}(
  swag_obj,
  min_dim_method = "median",
  min_dim_min_cv_error_quantile = 0.01
)
}
\arguments{
\item{swag_obj}{A \code{swag_obj}.}

\item{min_dim_method}{A \code{string} that specify the method to identify the dimension on which to compute the quantile to set the minimal CV to select model.}

\item{min_dim_min_cv_error_quantile}{The quantile of CV error in the selected dimension to specify the minimum CV value for selected models.}
}
\description{
return a formatted output of a \code{swag} object based on some model selection rules to be specified.
}
\author{
Gaetan Bakalli and Lionel Voirol
}
